refactor: move registration slice into core application
