+++ backend/src/persistence/application/use_cases/__init__.py
@@ -1,4 +1,5 @@
 from core.application.models import (
+    AdminRegistration,
     PenaAccountabilityExpenseCreate,
     PenaAccountabilityExpenseInfo,
     PenaAccountabilityInfo,
@@ -15,6 +16,9 @@ from core.application.models import (
     PenaProfileUpdate,
     PenasPage,
     PlayerProfile,
+    RegisteredAdmin,
+    RegisteredUser,
+    UserRegistration,
 )
 from core.application.models import PenaInfo as PenaSummary
 from core.application.use_cases.generate_pena_link_token_usecase import (
@@ -54,6 +58,23 @@ from core.application.use_cases.manage_pena_profile_usecase import (
     PenaProfileAccessDeniedError,
     PenaProfileNotFoundError,
 )
+from core.application.use_cases.register_admin_usecase import (
+    InvalidAdminRegistrationDataError,
+    RegisterAdminUseCase,
+)
+from core.application.use_cases.register_admin_usecase import (
+    UsernameAlreadyExistsError as AdminUsernameExistsError,
+)
+from core.application.use_cases.register_user_usecase import (
+    InvalidNationalityError as UserInvalidNationalityError,
+)
+from core.application.use_cases.register_user_usecase import (
+    InvalidRegistrationDataError,
+    RegisterUserUseCase,
+)
+from core.application.use_cases.register_user_usecase import (
+    UsernameAlreadyExistsError as UserUsernameExistsError,
+)
 from core.application.use_cases.update_player_profile_usecase import (
     InvalidNationalityError as PlayerInvalidNationalityError,
 )
@@ -87,27 +108,6 @@ from .manage_season_competition_usecase import (
 from .manage_season_lifecycle_usecase import ManageSeasonLifecycleUseCase
 from .manage_season_matches_usecase import ManageSeasonMatchesUseCase
 from .manage_season_players_usecase import ManageSeasonPlayersUseCase
-from .register_admin_usecase import (
-    AdminRegistration,
-    InvalidAdminRegistrationDataError,
-    RegisterAdminUseCase,
-    RegisteredAdmin,
-)
-from .register_admin_usecase import (
-    UsernameAlreadyExistsError as AdminUsernameExistsError,
-)
-from .register_user_usecase import (
-    InvalidNationalityError as UserInvalidNationalityError,
-)
-from .register_user_usecase import (
-    InvalidRegistrationDataError,
-    RegisteredUser,
-    RegisterUserUseCase,
-    UserRegistration,
-)
-from .register_user_usecase import (
-    UsernameAlreadyExistsError as UserUsernameExistsError,
-)
 from .season_competition_errors import (
     InvalidSeasonDataError,
     InvalidSeasonInsightsDataError,
