+++ backend/src/persistence/application/use_cases/register_admin_usecase.py
@@ -1,48 +1,14 @@
-from dataclasses import dataclass
-
-from auth.security import hash_password
-from persistence.application.ports.registration_port import (
-    AdminRegistrationPort,
-    DuplicateUsernameError,
+from core.application.models import AdminRegistration, RegisteredAdmin
+from core.application.use_cases.register_admin_usecase import (
+    InvalidAdminRegistrationDataError,
+    RegisterAdminUseCase,
+    UsernameAlreadyExistsError,
 )
 
-
-class UsernameAlreadyExistsError(Exception):
-    pass
-
-
-class InvalidAdminRegistrationDataError(Exception):
-    pass
-
-
-@dataclass(frozen=True)
-class AdminRegistration:
-    username: str
-    password: str
-    name: str
-
-
-@dataclass(frozen=True)
-class RegisteredAdmin:
-    admin_id: int
-    admin_guid: str
-
-
-class RegisterAdminUseCase:
-    def __init__(self, repository: AdminRegistrationPort):
-        self.repository = repository
-
-    def execute(self, data: AdminRegistration) -> RegisteredAdmin:
-        username = data.username.strip()
-        name = data.name.strip()
-        if not username or not name:
-            raise InvalidAdminRegistrationDataError()
-        try:
-            registered = self.repository.register_admin(
-                username=username,
-                password_hash=hash_password(data.password),
-                name=name,
-            )
-        except DuplicateUsernameError as exc:
-            raise UsernameAlreadyExistsError() from exc
-        return RegisteredAdmin(admin_id=registered.admin_id, admin_guid=registered.admin_guid)
+__all__ = [
+    "AdminRegistration",
+    "InvalidAdminRegistrationDataError",
+    "RegisterAdminUseCase",
+    "RegisteredAdmin",
+    "UsernameAlreadyExistsError",
+]
